+++ src/views/basics.mjs
@@ -161,7 +161,6 @@ export default async function (trie, theme) {
   const amount = null;
   const parser = JSON.parse;
   const startDatetime = null;
-  const accounts = await registry.accounts();
   const delegations = await registry.delegations();
   const href = null;
   const messages = await store.posts(
@@ -170,7 +169,6 @@ export default async function (trie, theme) {
     amount,
     parser,
     startDatetime,
-    accounts,
     delegations,
     href,
     "amplify",
@@ -181,7 +179,6 @@ export default async function (trie, theme) {
     amount,
     parser,
     startDatetime,
-    accounts,
     delegations,
     href,
     "comment",